requirements.txt
openai-whisper  # Whisper version
pyaudio          # For recording audio
numpy        # For audio data manipulation
